Refactor: extract `Triangle3d`, `TrianglePad` into `modmesh/universe/polygon.hpp`
