+++ cpp/modmesh/universe/CMakeLists.txt
@@ -7,6 +7,7 @@ set(MODMESH_UNIVERSE_HEADERS
     ${CMAKE_CURRENT_SOURCE_DIR}/bernstein.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/bezier.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/coord.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/polygon.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/World.hpp
     CACHE FILEPATH "" FORCE)
